+++ packages/disco/lib/src/disco_internal.dart
@@ -7,7 +7,6 @@ part 'models/overrides/override.dart';
 part 'models/providers/instantiable_provider.dart';
 part 'models/providers/arg_provider.dart';
 part 'models/providers/no_arg_provider.dart';
-part 'utils/disco_config.dart';
 part 'utils/extensions.dart';
 part 'widgets/provider_scope.dart';
 part 'widgets/provider_scope_override.dart';