+++ packages/disco/lib/src/utils/disco_config.dart
@@ -1,7 +1,0 @@
-part of '../disco_internal.dart';
-
-/// The global preferences for the Disco package.
-abstract final class DiscoConfig {
-  /// {@macro Provider.lazy}
-  static bool lazy = true;
-}